Move fake_stripe_id to tests/data
