+++ tests/unit/test_ingest_stripe.py
@@ -41,7 +41,8 @@ from ctms.schemas import (
     StripeSubscriptionCreateSchema,
     StripeSubscriptionItemCreateSchema,
 )
-from tests.unit.conftest import fake_stripe_id, unix_timestamp
+from tests.data import fake_stripe_id
+from tests.unit.conftest import unix_timestamp
 
 
 @pytest.fixture
